@@ -1,6 +1,6 @@
 {
   "name": "www",
-  "packageManager": "yarn@3.0.2",
+  "version": "0.0.0",
   "main": "index.js",
   "scripts": {
     "develop": "./gatsby_wrapper develop",
@@ -9,39 +9,39 @@
     "clean": "gatsby clean"
   },
   "dependencies": {
-    "@mdx-js/mdx": "v1",
-    "@mdx-js/react": "v1",
-    "@primer/react": "latest",
-    "@types/lz-string": "^1.3.34",
-    "assert": "^2.0.0",
-    "co_lc": "workspace:co_lc/js",
-    "cor": "workspace:cor/js",
+    "@mdx-js/mdx": "^3.0.0",
+    "@mdx-js/react": "^2.3.0",
+    "@primer/react": "^36.1.0",
+    "@types/react-helmet": "^6.1.9",
+    "@types/styled-system": "^5.1.22",
+    "assert": "^2.1.0",
+    "co_lc": "0.0.0",
+    "cor": "0.0.0",
     "events": "^3.3.0",
-    "ft": "workspace:ft/js",
-    "gatsby": "latest",
-    "gatsby-plugin-mdx": "^3.16.0",
-    "gatsby-source-filesystem": "latest",
-    "gtlc": "workspace:gtlc/js",
-    "ho21": "workspace:ho21/js",
-    "lz-string": "^1.4.4",
-    "monaco-editor": "^0.28.0",
-    "monaco-editor-webpack-plugin": "^4.1.2",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "ft": "0.0.0",
+    "gatsby": "^5.12.9",
+    "gatsby-plugin-mdx": "^5.12.3",
+    "gatsby-source-filesystem": "^5.12.1",
+    "gtlc": "0.0.0",
+    "ho21": "0.0.0",
+    "lz-string": "^1.5.0",
+    "monaco-editor": "^0.28.1",
+    "monaco-editor-webpack-plugin": "^4.2.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-helmet": "6.1.0",
     "react-markdown": "latest",
-    "styled-components": "latest",
-    "tal": "workspace:TAL/js",
-    "tiger": "workspace:tiger/js",
+    "styled-components": "^5.3.11",
+    "styled-system": "^5.1.5",
+    "tal": "0.0.0",
+    "tiger": "0.0.0",
     "typescript": "latest",
-    "www": "^0.0.0",
     "zlib": "^1.0.5"
   },
   "devDependencies": {
-    "@types/node": "^16.9.4",
-    "@types/react": "17.0.2",
-    "@types/react-dom": "17.0.2",
-    "@types/react-helmet": "6.1.0",
+    "@types/node": "^16.18.61",
+    "@types/react": "18.2.0",
+    "@types/react-dom": "18.2.0",
     "@types/styled-components": "latest",
     "gatsby-plugin-typescript": "latest"
   }
